Fix bug with loop indices not being mapped to int64 in C
This should fix the issue in cheri128

Also introduce a feature to more easily debug the C backend:

sail -dfunction Name

will pretty-print the ANF and IR representation of just the Name
function. I want to make this work for the type-checker as well, but
it's a bit hard to get that to not fire during re-writing passes right
now.