Canonicalise functions types in val specs
This brings Sail closer to MiniSail, and means that

type my_range 'n 'm = {'o, 'n <= 'o <= 'm. int('o)}

will work on the left hand side of a function type in the same way as
a regular built-in range type. This means that in principle neither
range nor int need be built-in types, as both can be implemented in
terms of int('n) (atom internally). It also means we can easily
identify type variables that need to be made into implict arguments,
with the criterion for that being simply any type variable that
doesn't appear in a base type on the LHS of the function, or only
appears on the RHS.