C: Fix miscompilation of constrained struct field access
For a Int-parameterised struct F('x: Int) = ... the optimizer would
attempt to optimize field access in cases where 'x was known to
constrain the types of the struct fields only locally. Which would
create a type error in the generated C. Now we always use the type
from the global struct type.

However, we previously weren't using struct type quantifiers to
optimize the field representation, which we now do.

Also rename some utility functions to better match the List
functions in the OCaml stdlib.