Mappings: Using ordinary patterns in forwards/backwards mappings
If we are preserving mappings for some backends, then patterns within
mappings must be re-written using the normal pattern rewrites, so we
have to use pexp rather than mpexp * exp for MCL_forwards and
MCL_backwards. This does mean the existing mapping realisation rewrite
no longer works. Still need to modify the rewriter so it actually
traverses into mappings because I don't think it is set up to do that
right now.