Separate enum members from identifiers in Jib IR
We need to reduce the SSA control flow graph before generating SMT for
concrete instructions because the unreachable states generate a lot of
read/write events. It'd be much nicer if this was possible without
having to use a type-checking environment to distinguish between
enumeration elements and enumeration constructors. We might also want
to separate V_id into V_id and V_register for ordinarly local
variables and registers too.