First go at using information from one side of a polymapping to check the other side
Previously this used a Typ_internal_unknown, but this was a bit ugly
because it was visible outside the type system. This version catches
pattern identifier inference errors in at most one side of the mapping
and re-checks that side with information from the other side of the
mapping.

This does require that at least one side is fully type-checkable,
which was not the case previously. A better approach might be to use
empty type annotations rather than a Typ_internal_unknown, but
otherwise follow Jon's previous method.

Also we can allow the mapping(mpat) syntax as a special case for
mpat <- mapping() which is both more concise and should allow the RISC-V
mappings to be ported over without difficulty