Start on generalising guard removal
Introduce a new internal node E_internal_cascade, which represents a
'cascade' of fallthroughs for a re-written pattern followed by a match
statement. The reason to introduce a new internal node is because:

1) The evaluation order for the fallthroughs needs to be lazy, which
Sail letbindings do not (and should not) support, this means they
should be translated to 'let fallthroughN () = ...' nested functions
in each backend. This is required to ensure that effects occur in the
correct order.

2) By using an internal node we can precisely control where
annotations must be added, and avoid duplicated type-checking, also it
makes it easy to ensure that if the original mapping always
type-checks, then the re-written cascade will too, as well as making
it much easier to construct a type preserving rewrite, as the scoping
and types of the fallthroughs is clear - if we tried to use existing
AST nodes to construct an equivalent expression, we might run into the
case where we need to insert annotations, but not all types are
neccessarily representable in the source if they contain generated
type variables, and this solution cleanly avoids those issues as the
typing rule for E_internal_cascade can just know about the type of the
fallthroughs in a way that E_let or similar could not.