Fix type check errors for scattered mapping uses before the final clause
This test was giving the following error:
    Type error:
    [test/c/scattered_mapping.sail]:12:23-44
    12 |function unit_str () = unit_str_map_forwards(())
       |                       ^-------------------^
       | No val spec found for unit_str_map_forwards
       |

This is because we were erroneously putting the realised val specs next
to their function definitions, which are put in the place of the final
clause (as a result of descattering), so once we strip the mapping val
spec itself, subsequent type checks will fail. Instead, put them in
place of the mapping val spec, or not at all if it does not exist. This
also removes the need for remove_mapping_valspecs, but requires us to
add a new rewrite in its place to strip out duplicate val specs, in case
any of the realised ones are already provided by the user.