Add new hex_bits_of_prefix function to C/OCaml/Lem libraries
This is intended to replace the hex_bits_N_matches_prefix functions,
providing a polymorphic version in its place. More importantly, it
removes the surrounding option type, as the C library is unable to
determine the right integer value to use for the option's kind, as it
varies based on what option type instantiations are used by the user's
input Sail. Instead, force the caller to turn it into an option type, if
desired, based on the length returned.