Mono: Use more environment information when adding bitvector casts
When considering whether to add a cast, now also consider the updated
environment within an if branch / match clause to compare against the
outer environment.  This picks up not only constraints on type variables
added by an if condition or pattern guard (e.g. "if (size == 16) ..."),
but also constraints depending on those (e.g. in "bits('width)" where
"'width == 'size * 8").  Fixes a type error observed when generating Lem
for sail-arm (in aget__Mem).