Add github actions to build on macOS and ubuntu
This commit adds two github action to build Sail on macOS and ubuntu (both using the latest version
of each for now). These just build and don't run any tests, as we run those on our own Jenkins
server which is much faster than the github build runners.

I also fixed INSTALL.md to include brew installing pkg-config on macOS as this seems to be required.

From testing on a personal fork it seems quite email happy when it fails. Maybe that's what we want
though.

There's also a windows option but I leave that as future work...