Handle more cases in bitvector cast rewrite
Add casts for function arguments using the constraints in the
environment of the function clause (not just assertions within the
function body).  Also pass in the global typing environment for
comparison with the environment within the function clause (and make a
corresponding change in the Lem pretty-printer so that it uses the right
environments).