Mono: Try to fix bug in inter-procedural analysis
The monomorphisation analysis decides whether to split function
arguments in the callee or in callers.  The code previously used a
datastructure that can hold results of either the one case or the other,
but there might be functions that are called in different contexts
leading to different decisions.  This patch changes the datastructure to
support storing all instances of either case.