Output INT64_MIN in code generator for min 64-bit integer literal
Fixes the warning generated because in C -X where X is the minimum integer is
parsed as a positive integer which is then negated. This causes a (I
believe spurious) warning that the integer literal is too large for the type.

Also using INT64_C so we get either long or long long depending on
platform