Docs: Mention "opam pin" and reword a detail
opam pin is useful in the development workflow.

About the rewording: The tests of Sail check the behavior of Sail and
not whether it is installed correctly, because the instructions above
that paragraph don’t install Sail, they just build it.
These circumstances weren’t represented appropriately by the text.