Strip leading *'s from saildoc
We now parse

/*!
 * Paragraph
 */

and

/*!
 *Paragraph
 */

the same as

/*!
Paragraph
*/

since the first form is prettier, and similar to what Doxygen, Javadoc
and other such tools allow. This can cause mild confusion, as if the
start of a line in the final form happens to have a * then it will
unexpectedly remove it, but this is a problem shared by those tools too
and the intent is that everyone just use the first form and never need
to worry about it.