Isabelle lib for cancel/sort/distinct
The fastest way to prove a list is distinct is to show it
is a permutation of a sorted version (sorted outside the
logic). In Isabelle-land, a permutation is proven by showing
the multisets are the same, which boils down to cancelling
common terms in an equality of additions. This theory provides
the required nuts and bolts.