Make -O no longer imply -no_effects
Fixes #122

This used to be needed as we required effect annotions to be precise.
Now annotions can specify a superset of the actual effects, in
particular for this change the `undef` effect for the non-determinism when
creating undefined values.

Tested with the current checkout of sail-riscv and appears to cause no
issues